Extact SetAliasesMethodCallExtractor and AttributeValueResolver
